installation scripts

+++ pythia/install_pythia.sh
@@ -4,6 +4,7 @@ PREFIX=$PWD
 BUILDDIR=$PWD
 PYTHIA_URL="https://github.com/libPythia/pythia.git"
 COMMIT_HASH=f353a754bcc1fe6c525a14f81ae3c4faf9967856
+FORCE=n
 
 
 usage()
@@ -15,9 +16,10 @@ OPTIONS:
    -h                               Show this message
    -p <director>		    Path to the installation directory
    -b <directory>	       	    Path to the build directory
+   -f 				    Force installation
 EOF
 }
-while getopts 'p:b:h' OPTION; do
+while getopts 'p:b:hf' OPTION; do
   case $OPTION in
   p)
 	PREFIX=$OPTARG
@@ -25,9 +27,13 @@ while getopts 'p:b:h' OPTION; do
   b)
 	BUILDDIR=$OPTARG
 	;;
-  h)	usage
-	exit 2
-	;;
+  f)
+      FORCE=y
+      ;;
+  h)
+      usage
+      exit 2
+      ;;
   esac
 done
 
@@ -37,15 +43,20 @@ shift $(($OPTIND - 1))
 # get Pythia
 mkdir -p "$BUILDDIR" 2>/dev/null
 cd "$BUILDDIR"
-if [ -d "pythia" ]; then
+if [ -d "pythia" ] && [ "$FORCE" = "y" ] ; then
     rm -rf "pythia"
 fi
-git clone "$PYTHIA_URL" pythia || exit 1
-cd pythia || exit 1
-git reset --hard $COMMIT_HASH || exit 1
 
-# install pythia
-meson build -Dprefix="$PREFIX" || exit 1
-cd build || exit 1
-ninja install || exit 1
+
+if ! [ -d "pythia" ]; then
+    git clone "$PYTHIA_URL" pythia || exit 1
+    cd pythia || exit 1
+    git reset --hard $COMMIT_HASH || exit 1
+
+    # install pythia
+    rm -rf build 2>/dev/null
+    meson build -Dprefix="$PREFIX" || exit 1
+    cd build || exit 1
+    ninja install || exit 1
+fi
